Add files via upload
Implemented fread and fwrite to read in all of the characters from a text file, and write them out into stdout, which can be redirected (as specified in the project desc.) into a zip file. Thus far, all the program does is copy the text from the file into a .z file, without compressing it. Implemented the beginning of the compression algorithm, where there is a check to see if the current character is the same as the next (using f read). 

Need to implement: multiple file handling, no file handling, compression algorithm.